fix: Correct import paths after restructuring
Fixed remaining broken import paths:

1. src/core/config/discovery.ts:
   - ../types/providers -> ../../types/providers
   (core/config needs to go up two levels to reach src/types)

2. src/providers/aws/provider.ts:
   - ../../analytics/anomaly-detector -> ../../core/analytics/anomaly
   (anomaly detector was moved to core/analytics/anomaly/ in restructuring)

These fix the last build errors in CI pipelines.
